fix behave errors
